requirements.txt
langchain==0.2.5
langchain-openai==0.1.7
langchain-community==0.2.5
python-dotenv>=1.0.0
numexpr>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

